requirements.txt
fastapi
uv
uvicorn
openai
python-dotenv
Pillow
faker
tesseract
pytesseract